// src/app/core/models/employee.model.ts
export interface Employee {
  employeeId: number;
  firstName: string;
  lastName: string;
  dateOfBirth: string;    // ISO format: YYYY-MM-DD
  address: string;
  gender: string;
  email: string;
  nation: string;
  phoneNumber: string;
  hireDate: string;       // ISO format: YYYY-MM-DD
  departmentId: number;
  positionId: number;
}